#672 Nested in new namespace of its own. Renamed File to FileRanges.
