+++ Assets/SEE/DataModel/DG/SourceRange.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 21b39035fe51e7649bf8e0d5e2470968
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
